Add files via upload
I followed a bunch of react tutorials, and I ended up coding the drive portion of the front end. But because I'm learning it as I'm going, the only things I could understand how to do with the time sensitivity was hard coding, so its 3000+ lines of code. I'm going to work on simplifying it into components and making it responsive but trying to learn all of this for the first time so bear w me! Uploading this so Elan can help me tomorrow fyi.